moved code into lottery.py, added waitlist table and admin view
moved lottery and user weight manipulation into lottery.py, added a waitlist table and admin view. admin view for waitlist has the option to move the next response in line (ranked 1) from on to off the waitlist, which updates the rest of the waitlists rankings to one above their previous rank. places that need more code are marked ACTION REQUIRED